add constants file to all apps
